fix compilation errors

+++ include/kiwi/ScriptType.h
@@ -246,5 +246,5 @@ namespace kiwi
      * 
      * @return 0 if the character is not an emoji, 1 if c0 is an emoji, 2 if c0 and c1 are combined to form an emoji.
      */
-    size_t isEmoji(char32_t c0, char32_t c1 = 0);
+    int isEmoji(char32_t c0, char32_t c1 = 0);
 }
